Fix error handling for subscriptions.
The `error` function for `$.ajax` just receives a `jqXHR` object
instead of parsed data. We return an error code in cases where the
card has been declined, which triggers the error handler instead of
the success handler. So we need a dedicated error handler instead
of trying to overload the success handler to handle all the cases.