Ensure subscription change errors are shown to the user.
Calling `valid?` will clear any errors that have been stored against
the `Payola::Subscription` instance, which means any errors from
Stripe that were stored as part of the service objects will be lost
and the user will always be told that the change was successful.

The solution is to avoid calling `valid?`, and the simplest approach
to see if there actually _weren't_ any errors is to use
`errors.empty?`.