Fix jQuery XHR JSON error handling
Annoyingly, jQuery doesn't parse the response as JSON if the status code
is an error (http://stackoverflow.com/q/12310702/117104), even though
the server response's Content-Type is correctly set to application/json.
So in our XHR error handlers, xhr.responseJSON was always undefined.

Manually parse the responseText as JSON using jQuery.parseJSON().

Note that this won't behave correctly if the server doesn't return JSON
(eg. if there's a 500 error for some reason and it returns an HTML error
page or something, or there's a 404 and it renders 'nothing').It would
be good to handle those cases. I think we'd want to DRY up the error
handlers first though as it's a bit verbose to handle all these cases
correctly.