Pass at_period_end param to CancelSubscription, support it in subscriptions/_cancel partial
We already have support in CancelSubscription for the at_period_end
param, but you had to call that service manually to be able to pass that
argument through.

Add support for an at_period_end param in the stock
subscriptions/_cancel partial, and pass it through to the
CancelSubscription service from the SubscriptionsController#destroy
method.