Update StartSubscription service to take an optional stripe_customer_id
We currently have support for reusing a Stripe customer based on the
subscription owner having an existing successful Subscription stored in
the database.

Extend this behaviour to allow an explicit stripe_customer_id to be
passed to StartSubscription.call().

This is useful if you have Stripe customers that have made one-off purchases
of some variety, and want to reuse the same Stripe customer when they choose
to start a subscription.