Don't set a stripe_customer_id in subscription factory by default
Now if a subscription has a stripe_customer_id present when it is
created it is treated differently - we attempt to look up and use
the existing Stripe customer. In our specs, we normally don't want this
to be the case. We only want a stripe_customer_id to be present when
testing the new functionality, or when we're dealing with a subscription
that's already active (eg. when updating it or cancelling it).

Stop setting the stripe_customer_id by default in the factory, and set
it explicitly when we expect it to be present.