Fix StartSubscription spec that was incidentally passing, then broke with my stripe_customer_id changes
This spec used to pass because creating the subscription from the
factory was already assigning it a stripe_customer_id that wasn't nil,
and wasn't the same as the first subscription we created. The intent of
the spec was that a new stripe customer would be created, and the second
subscription would be associated with that stripe_customer_id, but that
wasn't occurring. Instead, the second subscription was ending up in an
errored state, because the stripe token generated by the factory wasn't
a valid token.

My new stripe_customer_id handling revealed that this spec wasn't
behaving as intended. This commit changes the spec to ensure that the
second subscription fails, as it is created without a stripe token, and
the existing customer id we were able to find has already been deleted.