Update cancel_at_period_end in CancelSubscription service
If CancelSubscription is called with `at_period_end: true`, also update
the subscription to store that it will be canceled at the end of the
billing period.

Previously, this attribute wouldn't be changed until the Payola
subscription was synced with the associated Stripe subscription (ie.
when a webhook arrived).

Also, add a test to assert that the Payola subscription isn't
immediately canceled if the `at_period_end: true` option is used.