Maintain the active coupon code on a subscription
Stripe removes coupon codes from a subscription after their repeating
duration has elapsed. When this happens, or when the subscription is
changed for any other reason, we want to update the coupon store on the
Payola subscription.

We need to check whether the stripe_sub will `respond_to?(:discount)`
before attempting to copy the value because this attribute is added to
stripe-ruby-mock in version 2.2.0.