Disallow deleting a plan if it has any related subscriptions
Even if Stripe allows deleting a plan with subscriptions (and it will
keep the subscriptions), for referential integrity purposes in the
relational DB we shouldn't allow it. If you want to allow deleting a
plan then you should implement some variation of soft-deletes. I'm sure
that's what Stripe really does in the background so "deleting a plan"
actually just hides it.